ba/requirements.txt
click==8.1.3
distro==1.7.0
Flask==2.1.2
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.0
pandas==1.4.3
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
tabula-py==2.4.0
Werkzeug==2.1.2
zipp==3.8.0
